third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/background-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'background' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'background' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'background' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'background' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'background' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
Harness: the test ran to completion.
